Move tests to subpackage
